Fixed annotation deletion sync issue, quality of life improvements
This commit fixes the annotation deletion issue. Now it is possible to
delete both in Zotero or KOreader and annotations will be deleted.

It includes some minor enhancements and bug fixes such as using the
network manager for syncing to make sure WiFi is connected and it also
fixes a color sync issue

Also formatting and overall cleanup of the code such as adding the
Zotero: to all debug messages and removing some unused functions and
imports